fix(dracut-systemd): do not try to remove mount hooks in read-only locations
Commit 04d5e2944e1b1a85a945910995cb8df00d983a7c introduced read-only locations
for hooks. After a mount hook script located in these locations is executed, if
`$NEWROOT` is mounted but `usable_root()` returns false, `rm` will print a
permission denied message.

Follow-up for 04d5e2944e1b1a85a945910995cb8df00d983a7c

+++ modules.d/77dracut-systemd/dracut-mount.sh
@@ -23,9 +23,14 @@ while :; do
         . "$f"
         if ismounted "$NEWROOT"; then
             usable_root "$NEWROOT" && break
-            warn "$NEWROOT has no proper rootfs layout, ignoring and removing offending mount hook"
             umount "$NEWROOT"
-            rm -f -- "$f"
+            # hooks can be in read-only locations
+            if [ -w "$f" ]; then
+                warn "$NEWROOT has no proper rootfs layout, ignoring and removing offending mount hook"
+                rm -f -- "$f"
+            else
+                warn "$NEWROOT has no proper rootfs layout, but the offending mount hook is on a read-only location and cannot be removed"
+            fi
         fi
     done
